pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "procopt"
version = "0.1.0"
description = "Process Optimization Tool"
authors = [
    {name = "Michael Wornow", email = "mwornow@stanford.edu"},
]
requires-python = ">=3.10"
dependencies = [
    "Flask==3.0.2",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-CORS==4.0.0",
    "pika==1.3.2",
    "python-dotenv==1.0.1",
    "Werkzeug==3.0.1",
    "Pillow==10.2.0",
    "PyMuPDF==1.23.26",
    "pydantic==2.10.6",
    "annotated-types==0.7.0",
    "pydantic-core==2.27.2",
    "numpy==2.2.3",
    "litellm>=1.30.3",
    "openai>=1.0.0"
]

[tool.setuptools.packages.find]
include = ["procopt*"]
exclude = ["ignore*"]